endpoints extractions: use the converter func instead of the one from util
