+++ ec/internal/util/helpers.go
@@ -18,7 +18,6 @@
 package util
 
 import (
-	"fmt"
 	"os"
 	"strconv"
 
@@ -30,26 +29,6 @@ import (
 
 // used in tests
 var GetEnv = os.Getenv
-
-// FlattenClusterEndpoint receives a ClusterMetadataInfo, parses the http and
-// https endpoints and returns a map with two keys: `http_endpoint` and
-// `https_endpoint`
-func FlattenClusterEndpoint(metadata *models.ClusterMetadataInfo) map[string]interface{} {
-	if metadata == nil || metadata.Endpoint == "" || metadata.Ports == nil {
-		return nil
-	}
-
-	var m = make(map[string]interface{})
-	if metadata.Ports.HTTP != nil {
-		m["http_endpoint"] = fmt.Sprintf("http://%s:%d", metadata.Endpoint, *metadata.Ports.HTTP)
-	}
-
-	if metadata.Ports.HTTPS != nil {
-		m["https_endpoint"] = fmt.Sprintf("https://%s:%d", metadata.Endpoint, *metadata.Ports.HTTPS)
-	}
-
-	return m
-}
 
 // IsCurrentEsPlanEmpty checks that the elasticsearch resource current plan is empty.
 func IsCurrentEsPlanEmpty(res *models.ElasticsearchResourceInfo) bool {
